PullRequest: graal/10540

+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/graal/nodes/NewStoredContinuationNode.java
@@ -49,5 +49,5 @@ public class NewStoredContinuationNode extends AbstractNewObjectNode implements 
     }
 
     @NodeIntrinsic
-    public static native StoredContinuation allocate(long size);
+    public static native StoredContinuation allocate(int size);
 }
